nagios-plugins/requirements.txt
python-numpy
python-requests
python-BeautifulSoup
python-simplejson
pynag

